remove routing package and reimplement old routing style #13
