+++ packages/frontend/package.json
@@ -26,7 +26,6 @@
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "elite-feature-flags": "^1.0.0",
-    "elite-configuration": "^1.0.0",
-    "elite-routing": "^1.0.0"
+    "elite-configuration": "^1.0.0"
   }
 }
