+++ packages/frontend/src/components/App.tsx
@@ -1,27 +1,25 @@
-import * as React from 'react';
-import { hot } from 'react-hot-loader';
-import { Route, Switch, Redirect } from 'react-router';
-import { Router } from 'react-router-dom';
-import history from '../util/history';
+import { getConfiguration } from 'elite-configuration';
 import { FeatureFlagsProvider } from 'elite-feature-flags';
 import { Configuration } from 'elite-types';
-import { getConfiguration } from 'elite-configuration';
-import { getAllRegisteredAppRoutes } from 'elite-routing';
-import { AppPaths, registerRoutes } from '../util/routes';
+import * as React from 'react';
+import { hot } from 'react-hot-loader';
+import { Redirect, Route, Switch } from 'react-router';
+import { Router } from 'react-router-dom';
+import history from '../util/history';
+import { AppPath, APP_ROUTES } from '../util/routes';
 
 // Global bootstrap: install subsystems and load configuration
-registerRoutes();
 const configuration: Configuration = getConfiguration();
 
 export const AppComponent = () => (
   <FeatureFlagsProvider value={configuration.featureMap}>
     <Router history={history}>
       <Switch>
-        {getAllRegisteredAppRoutes().map((routeProps, index) => (
+        {APP_ROUTES.map((routeProps, index) => (
           <Route key={index} {...routeProps} />
         ))}
         {/* Error 404 Fallback */}
-        <Redirect to={AppPaths.HOME} />
+        <Redirect to={AppPath.HOME} />
       </Switch>
     </Router>
   </FeatureFlagsProvider>
