+++ packages/frontend/src/components/pages/HomePage.tsx
@@ -1,8 +1,8 @@
+import { Divider } from '@material-ui/core';
+import { FeatureFlag } from 'elite-feature-flags';
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { LinkDirectory } from './support/LinkDirectory';
-import { Divider } from '@material-ui/core';
-import { FeatureFlag } from 'elite-feature-flags';
 
 export interface HomePageProps extends RouteComponentProps {}
