+++ packages/frontend/tsconfig.json
@@ -14,9 +14,6 @@
       "path": "../feature-flags"
     },
     {
-      "path": "../routing"
-    },
-    {
       "path": "../types"
     }
   ]
