--- packages/routing/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "elite-routing",
-  "version": "1.0.0",
-  "private": true,
-  "publishConfig": {
-    "access": "public"
-  },
-  "types": "dist/index.d.ts",
-  "main": "dist/index.js",
-  "scripts": {
-    "clean": "rm -rf dist/ node_modules/ tsconfig.tsbuildinfo"
-  },
-  "devDependencies": {
-    "@types/react": "^16.9.11",
-    "@types/react-router": "^5.1.3",
-    "elite-types": "^1.0.0"
-  },
-  "dependencies": {
-    "react": "^16.12.0"
-  }
-}
